make AIO pin look-up more flexible
The current approach used to look up the pin for an AIO channel
assumes that the AIO pins are positioned sequentially in the pins
array immediately after the GPIO pins.  This works for Arduino-style
boards, but doesn't port well to other boards where ADC pins may be
scattered around the pin array and providing GPIO functionality as
well.

This commit introduces a different approach by instead searching
the pin array for ADC-capable pins, and infers their AIO channel
numbers from the order in which they appear in the pin array.

This should still work for boards such as Galileo and Edison, and
while also supporting boards such as the UP board where the ADC
pin appears at an arbitrary position in the pin array.

Signed-off-by: Dan O'Donovan <dan@emutex.com>