rpi: add uart device_path
Currently there is no device_path registered for UART in Raspberry Pi
source. This will hinder the working of UART driver in mraa.

So, register the device_path with '/dev/ttyS0' for all the raspberry
pi models except Rpi3. For Rpi3 'dev/serial0' is used as most of the
raspbian images make this as a symlink to _real_ UART port by
default.

Tested-on: Raspberry pi 3 model B

Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>