python2/CMakeLists.txt: make SWIG doc interfaces depend on wrapper gen
Previously we used SWIG_MODULE_<name>_REAL_NAME to have the doc interfaces
depend on. However that target is preceeded by the SWIG generation in our case
gen_python2-mraa_mraa2PYTHON_wrap. Typically this is hidden because a parralel
build has so many other deps this can be hidden but this is the correct path.

Fixes #911

Signed-off-by: Brendan Le Foll <brendan@fridu.org>