mraa: fix mraa_file_contains() and mraa_file_contains_both()
These two functions may return wrong results if lines in the
file contains embedded '\0' characters.

For example, file "/proc/device-tree/compatible" could have
multiple values in one line (corresponding to the dts), and the
values are separated with '\0'.

In this situation, we need to replace the embedded '\0' and check
the entire line. If we don't do this, only the first value is
checked by strstr(), and the function may return false even if
the file contains certain content.

Signed-off-by: Benxi Liu <bxliu@linux.alibaba.com>