Add winston as explicit dependency for mraaStub
The `mraaStub` package uses `winston` for logging, but it doesn't explicitly depend on it. This was breaking our build. This being node, the obvious workaround is to add `winston` as a dependency in our own `package.json`. We've done this for now, but we don't actually use it in our own code. This is obviously a bug.

I'm aware that I've specified a really old version of `winston`, but it's one that I'm 100% sure works with `mraaStub`. If anyone's keen on testing what the latest compatible version is, that would of course be preferable, but in the mean time be advised that this one works fine